requirements.txt
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.2
openai==1.54.0
python-dotenv==1.0.1
